update taskNew options, questions
